refactor: dissolve node_tools module
MCP handler factory moves into mcp/tool.rs (schema + handler together).
Delegate becomes a top-level node module. No artificial node_tools
grouping.
