+++ crates/node/src/lib.rs
@@ -1,16 +1,14 @@
 //! Crabtalk node — runtime + transport + event loop.
 
 pub mod cron;
+pub mod delegate;
 pub mod event_bus;
 pub mod hook;
 pub mod mcp;
 pub mod node;
-pub mod node_tools;
 pub mod provider;
 #[cfg(feature = "fs")]
 pub mod storage;
-
-pub use tools::Memory;
 
 pub use hook::NodeEnv;
 #[cfg(unix)]
@@ -21,4 +19,5 @@ pub use node::{
     event::{NodeEvent, NodeEventSender},
     setup_tcp,
 };
+pub use tools::Memory;
 pub use wcore::NodeConfig;
