+++ crates/node/src/mcp/tool.rs
@@ -1,8 +1,14 @@
-//! Tool schema for the MCP meta-tool.
+//! MCP tool schema and handler factory.
 
+use super::McpHandler;
+use runtime::AgentScope;
 use schemars::JsonSchema;
 use serde::Deserialize;
-use wcore::agent::{AsTool, ToolDescription};
+use std::{
+    collections::BTreeMap,
+    sync::{Arc, RwLock},
+};
+use wcore::{ToolDispatch, ToolHandler, agent::ToolDescription};
 
 #[derive(Deserialize, JsonSchema)]
 pub struct Mcp {
@@ -19,6 +25,23 @@ impl ToolDescription for Mcp {
         "Call an MCP tool by name, or list available tools if no exact match.";
 }
 
-pub fn tools() -> Vec<wcore::model::Tool> {
-    vec![Mcp::as_tool()]
+/// Build a handler that dispatches MCP tool calls through the McpHandler.
+pub fn handler(
+    mcp: Arc<McpHandler>,
+    scopes: Arc<RwLock<BTreeMap<String, AgentScope>>>,
+) -> ToolHandler {
+    Arc::new(move |call: ToolDispatch| {
+        let mcp = mcp.clone();
+        let scopes = scopes.clone();
+        Box::pin(async move {
+            let allowed_mcps: Vec<String> = scopes
+                .read()
+                .expect("scopes lock poisoned")
+                .get(&call.agent)
+                .filter(|s| !s.mcps.is_empty())
+                .map(|s| s.mcps.clone())
+                .unwrap_or_default();
+            super::dispatch::dispatch_mcp(&mcp, &call.args, &allowed_mcps).await
+        })
+    })
 }
